Moved the deliverables_template_content to the general conftest
